+++ changelog.md
@@ -6,6 +6,8 @@
   ([#127](https://github.com/feltcoop/gro/pull/127))
 - add and publish `ui/PendingButton.svelte`
   ([#125](https://github.com/feltcoop/felt/pull/125))
+- add and publish `ui/FeltWindowHost.svelte` and `ui/FeltWindowGuest.svelte`
+  ([#128](https://github.com/feltcoop/felt/pull/128))
 
 ## 0.8.0